Plugin and database are now working with client
Need to figure out instance ID and ODL SFC logic
Implement create_wait function for ODL driver
